+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added automatic fetching of the system username for use in CronTab so that we do not
+have to ask the user for it and allow for automatic cron job scheduling.
+- Added functionality that allows user to schedule, start, or stop maintenance on their Minecraft server or hosting system.
 - The ability to run a user's own custom shell scripts placed in `./scripts/custom-command.sh`
 - Discord Integration. The user now has the ability to configure settings to post messages into their Discord server.
   - Unhandled exceptions. When the system runs into an unhandled error it will now post a message to your configured Discord server.
@@ -28,6 +31,7 @@ seconds before rebooting.
 
 ### Fixed
 - The temperature monitor was starting even when it hadn't been configured.
+- An issue with the Emailer preventing it from sending emails as expected
 
 ### Security